Chromosome Blueprint

API: ChromosomeBlueprint

Exported functions

Making an empty chromosome blueprint

Pseudoseq.plan_chrom — Function.

Create an empty blueprint for n copies of a chromosome of len base pairs.

source

Planning motif repetitions along chromosomes

Pseudoseq.plan_repetition — Function.

plan_repetition(cb::ChromosomeBlueprint, from::UnitRange{Int}, to::UnitRange{Int})

Plan a repetition in a chromosome, where the bases in the from region of the chromosome, are guaranteed to occur again in the to region of the chromosome.

Every copy of the chromosome will have this same repetition.

Creates a new chromosome blueprint, based on the input blueprint cb.

Note

In the new blueprint, the from and to regions of the planned chromosome will be consumed, and cannot be used to plan any other subsequently added features.

(See also: plan_het)

source

plan_repetition(cb::ChromosomeBlueprint, from::Int, to::Int, size::Int)

Plan a repetition in a chromosome, where the bases in the from:(from + size - 1) region of the chromosome, are guaranteed to occur again in the to:(to + size - 1) region of the chromosome.

Every copy of the chromosome will have this same repetition.

Creates a new chromosome blueprint, based on the input blueprint cb.

Note

In the new blueprint, the from and to regions of the planned chromosome will be consumed, and cannot be used to plan any other subsequently added features.

(See also: plan_het)

source

plan_repetition(cb::ChromosomeBlueprint, intervals::Vector{UnitRange{Int}})

Plan a series of repetitions in a chromosome.

Every copy of the chromosome will have these same repetitions.

Creates a new chromosome blueprint, based on the input blueprint cb.

Tip

Use the suggest_regions function to help decide on a set of sites to make heterozygous.

Note

The number of intervals provided must be an even number. This is because intervals 1 & 2 define the first repeat, intervals 3 & 4 define the second, and so on.

Note

In the new blueprint, the regions that were use to plan the repetitions, have been consumed, and cannot be used to plan any other subsequently added features.

(See also: plan_het)

source

Planning heterozygosity between chromosome copies

Pseudoseq.plan_het — Function.

plan_het(cb::ChromosomeBlueprint, pos::Int, alleles::Vector{DNA})

Plan heterozygosity between copies of a chromosome at position pos.

The alleles vector must contain a nucleotide for each copy of the chromosome.

For example if you provided a vector of [DNA_A, DNA_C], for the heterozygous site you're defining at pos, the first copy of the chromosome will have an A, and the second copy of the chromosome will have a C.

Creates a new chromosome blueprint, based on the input blueprint cb.

Tip

Use the suggest_alleles function to help decide on a set of alleles to use.

Note

In the new blueprint, the positions that were use to plan the heterozygosity, have been consumed, and cannot be used to plan any other subsequently added features.

(See also: plan_repetition)

source

plan_het(cb::ChromosomeBlueprint, pos, alleles::Vector{Vector{DNA}})

Plan heterozygosity between copies of a chromosome at the positions defined by pos.

The alleles vector must contain a vector for every position in pos that defines the base each copy of the chromosome has.

Creates a new chromosome blueprint, based on the input blueprint cb.

Tip

Use the suggest_alleles function to help decide on a set of alleles to use.

Tip

Use the suggest_regions function to help decide on a set of sites to make heterozygous.

Note

In the new blueprint, the positions that were use to plan the heterozygosity, have been consumed, and cannot be used to plan any other subsequently added features.

(See also: plan_repetition)

source

plan_het(cb::ChromosomeBlueprint, pos, nalleles::Int)

Plan heterozygosity between copies of a chromosome at the positions defined by pos.

This particular method of plan_het will generate a set of alleles for each position in pos for you automatically. You just have to specify how many possible alleles (nalleles) there are.

Creates a new chromosome blueprint, based on the input blueprint cb.

Tip

Use the suggest_regions function to help decide on a set of sites to make heterozygous.

Note

The maximum value nalleles can be is the number of chromosome copies your blueprint contains. If you try to give nalleles a higher number, then this method will behave as if nalleles is equal to the number of chromosome copies in the blueprint. This is because if you have say 2 copies of a chromosome, then those two copies can carry at maximum 2 different bases at a position.

Note

In the new blueprint, the positions that were use to plan the heterozygosity, have been consumed, and cannot be used to plan any other subsequently added features.

source

plan_het(cb::ChromosomeBlueprint, npos::Int, nalleles::Int)

Plan heterozygosity between copies of a chromosome at npos positions, each heterozygous positions will have nalleles distict alleles.

This particular method of plan_het will decide on a set of available positions along the chromosome in the blueprint. For each postition it will then decide on a set of nalleles alleles that will be applied to each copy of the chromosome.

Creates a new chromosome blueprint, based on the input blueprint cb.

Note

The maximum value nalleles can be is the number of chromosome copies your blueprint contains. If you try to give nalleles a higher number, then this method will behave as if nalleles is equal to the number of chromosome copies in the blueprint. This is because if you have say 2 copies of a chromosome, then those two copies can carry at maximum 2 different bases at a position.

Note

In the new blueprint, the positions that were use to plan the heterozygosity, have been consumed, and cannot be used to plan any other subsequently added features.

source