API: sequence

Exported functions

Pseudoseq.sequence — Function

sequence(input, output = nothing; kwargs...)

Run a sequencing experiment from start to finish.

This method is the highest-level function for running sequencing simulations with Pseudoseq. It runs the Pseudoseq sequencin worksflow described in the manual, tuned by the following keyword parameters:

Global parameters:

input: A filename or FASTA.Reader providing the input genome.

output: A filename or FASTA.Writer providing a destination for the output reads.

Parameters for workflow step 1 (Create a pool of DNA molecules):

ng: Integer; the number of genomes the molecule pool is initialized with.

Parameters for step 2 (Processing the DNA molecule pool):

tusize: Integer; the number of possible tags a DNA molecule may be tagged with (default: 0).

taggedflen: Integer; the desired average fragment length at which molecules are tagged (default: 0bp).

flen: Integer; the desired average fragment length at which molecules are sampled (default: 700bp).

cov: Integer; the desired expected coverage (default: 30x coverage).

Parameters for step 3 (Generating reads):

paired: true or false; Whether or not to sequence from both end of each molecule (default: true).

rdlen: Integer; the desired read length (default: 250bp).

err: Float; the desired per-base error rate (default: 0.001).