Chromosome Blueprint

API: Build-a-Genome

Exported functions

Making an empty chromosome blueprint

Pseudoseq.plan_chrom — Function.

Create an empty blueprint for n copies of a chromosome of len base pairs.

source

Planning motif repetitions along chromosomes

Pseudoseq.plan_repetition — Function.

plan_repetition(cb::ChromosomeBlueprint, from::UnitRange{Int}, to::UnitRange{Int})

Plan a repetition in a chromosome, where the bases in the from region of the chromosome, are guaranteed to occur again in the to region of the chromosome.

Every copy of the chromosome will have this same repetition.

Creates a new chromosome blueprint, based on the input blueprint cb.

Note

In the new blueprint, the from and to regions of the planned chromosome will be consumed, and cannot be used to plan any other subsequently added features.

source

plan_repetition(cb::ChromosomeBlueprint, from::Int, to::Int, size::Int)

Plan a repetition in a chromosome, where the bases in the from:(from + size - 1) region of the chromosome, are guaranteed to occur again in the to:(to + size - 1) region of the chromosome.

Every copy of the chromosome will have this same repetition.

Creates a new chromosome blueprint, based on the input blueprint cb.

source

plan_repetition(cb::ChromosomeBlueprint, intervals::Vector{UnitRange{Int}})

A conveinience method of plan_repetition. Designed to ease the process of planing a series of repetitions in a chromosome.

Every copy of the chromosome will have these same repetitions.

Creates a new chromosome blueprint, based on the input blueprint cb.

Tip

Use the suggest_regions function to help decide on a set of sites to make heterozygous.

Note

The number of intervals provided must be an even number. This is because intervals 1 & 2 define the first repeat, intervals 3 & 4 define the second, and so on.

Note

In the new blueprint, the regions that were use to plan the repetitions, have been consumed, and cannot be used to plan any other subsequently added features.

source

Planning heterozygosity between chromosome copies

Pseudoseq.plan_het — Function.

plan_het(cb::ChromosomeBlueprint, pos::Int, alleles::Vector{DNA})

Plan heterozygosity between copies of a chromosome at position pos.

The alleles vector must contain a nucleotide for each copy of the chromosome.

For example if you provided a vector of [DNA_A, DNA_C], for the heterozygous site you're defining at pos, the first copy of the chromosome will have an A, and the second copy of the chromosome will have a C.

Creates a new chromosome blueprint, based on the input blueprint cb.

Tip

Use the suggest_alleles function to help decide on a set of alleles to use.

Note

In the new blueprint, the positions that were used to plan the heterozygosity, have been consumed, and cannot be used to plan any other subsequently added features.

source

plan_het(cb::ChromosomeBlueprint, pos, alleles::Vector{Vector{DNA}})

A conveinience method of plan_het. Plans heterozygosity between copies of a chromosome at the many positions defined by pos.

The alleles vector must contain a vector for every position in pos that defines the base each copy of the chromosome has.

Creates a new chromosome blueprint, based on the input blueprint cb.

Tip

Use the suggest_alleles function to help decide on a set of alleles to use.

Tip

Use the suggest_regions function to help decide on a set of sites to make heterozygous.

Note

In the new blueprint, the positions that were use to plan the heterozygosity, have been consumed, and cannot be used to plan any other subsequently added features.

(See also: plan_repetition)

source

Fabricate

fabricate

Utility functions

Pseudoseq.suggest_regions — Function.

suggest_regions(cp::ChromosomeBlueprint, size::Int, n::Int)

A useful utility function to assist planning features in a chromosome blueprint.

This method returns a vector of non-overlapping, regions of the chromosome planned represented by the ChromosomeBlueprint cb.

These regions are free regions: they are untouched by any other planned features, and so may be used when planning other features (See also: plan_repetition, plan_het).

The regions will be sizebp in length.

Warn

This function was designed for use interactively in a julia session.

If this method cannot find n free regions of the size you've asked for, it will still give an output vector containing the regions it did manage to find, but it will issue a warning to the terminal. This will get increasingly likely as you fill the chromosome blueprint up with features.

If you are using this method in a script or program where you depend on a reliable number of regions, either add a check to make sure you got the number of regions you need, or use this method interactively, and hard code an appropriate output into your script.

source