Reads

API: Reads

Exported

Pseudoseq.make_reads — Function.

make_reads(::Type{PairedEnd}, u::Universe, flen::Int, rlen::Int = flen)

Create a set of paired-end reads from a universe of DNA molecules u.

flen sets the length of forward read, and rlen sets the length of the reverse read. If you only provide flen, then the function sets rlen = flen.

Note

If a molecule in the universe is not long enough to create a forward and/or reverse read, then that molecule will simply be skipped.

source